deduplication code refactor
