+++ openg2p_registration/models/__init__.py
@@ -10,3 +10,4 @@ from . import openg2p_registration_widget
 from . import openg2p_submission_registration_map
 from . import openg2p_registration_org_map
 from . import openg2p_registration_api
+from . import openg2p_deduplication
